+++ backend/src/main/java/com/daedan/festabook/lostitem/controller/LostItemController.java
@@ -1,6 +1,9 @@
 package com.daedan.festabook.lostitem.controller;
 
 import com.daedan.festabook.global.argumentresolver.OrganizationId;
+import com.daedan.festabook.lostitem.dto.LostItemRequest;
+import com.daedan.festabook.lostitem.dto.LostItemResponse;
+import com.daedan.festabook.lostitem.dto.LostItemResponses;
 import com.daedan.festabook.lostitem.service.LostItemService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
